+++ test/unit/test_ami_stream.cpp
@@ -133,7 +133,7 @@ int main(int argc, char **argv) {
 			}
 
 			tpie::stream_offset_type newoff = s.tell();
-			if (newoff != idx+1) {
+			if (static_cast<size_t>(newoff) != idx+1) {
 				std::cout << "Offset advanced to " << newoff << ", expected " << (idx+1) << std::endl;
 				return EXIT_FAILURE;
 			}
